requirements.txt
beautifulsoup4
langchain
langchain_google_genai
pandas
python-dotenv
Requests
selenium
tqdm
webdriver_manager
